+++ src/ScottPlot5/ScottPlot5/Interfaces/IPathStrategy.cs
@@ -1,0 +1,9 @@
+﻿namespace ScottPlot;
+
+/// <summary>
+/// Strategy for generating a path that connects a collection of pixels
+/// </summary>
+public interface IPathStrategy
+{
+    public SKPath GetPath(IEnumerable<Pixel> pixels);
+}