+++ src/ScottPlot5/ScottPlot5/Plottables/Scatter.cs
@@ -21,9 +21,22 @@ public class Scatter(IScatterSource data) : IPlottable
     public ConnectStyle ConnectStyle = ConnectStyle.Straight;
 
     /// <summary>
-    /// If enabled, points will be connected by smooth lines instead of straight lines.
+    /// Controls whether points are connected by smooth or straight lines
     /// </summary>
-    public bool Smooth = false;
+    public bool Smooth
+    {
+        set
+        {
+            PathStrategy = value
+                ? new PathStrategies.QuadHalfPoint()
+                : new PathStrategies.Straight();
+        }
+    }
+
+    /// <summary>
+    /// Strategy to use for generating the path used to connect points
+    /// </summary>
+    public IPathStrategy PathStrategy { get; set; } = new PathStrategies.Straight();
 
     public Color Color
     {
@@ -42,7 +55,7 @@ public class Scatter(IScatterSource data) : IPlottable
 
     public void Render(RenderPack rp)
     {
-        // TODO: can this be more effecient by moving this logic into the DataSource to avoid copying?
+        // TODO: can this be more efficient by moving this logic into the DataSource to avoid copying?
         Pixel[] markerPixels = Data.GetScatterPoints().Select(Axes.GetPixel).ToArray();
 
         if (!markerPixels.Any())
@@ -57,21 +70,14 @@ public class Scatter(IScatterSource data) : IPlottable
         };
 
         using SKPaint paint = new();
+        using SKPath path = PathStrategy.GetPath(linePixels);
 
-        if (Smooth)
-        {
-            Drawing.DrawSmoothLines1(rp.Canvas, paint, linePixels, LineStyle);
-        }
-        else
-        {
-            Drawing.DrawLines(rp.Canvas, paint, linePixels, LineStyle);
-        }
-
+        Drawing.DrawLines(rp.Canvas, paint, path, LineStyle);
         Drawing.DrawMarkers(rp.Canvas, paint, markerPixels, MarkerStyle);
     }
 
     /// <summary>
-    /// Convert scatter plot points (connected by diagnal lines) to step plot points (connected by right angles)
+    /// Convert scatter plot points (connected by diagonal lines) to step plot points (connected by right angles)
     /// by inserting an extra point between each of the original data points to result in L-shaped steps.
     /// </summary>
     /// <param name="points">Array of corner positions</param>